backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.8.0
bcrypt==4.1.1
cryptography>=41.0.0

# Validation
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# OpenAI
openai>=1.0.0

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# Email
aiosmtplib==3.0.1
aioimaplib==1.0.1

# Templating (for HTML emails)
jinja2==3.1.2

# PDF generation
weasyprint==60.1

# Utilities
python-dateutil==2.8.2
